refactor(documents): use feature modules for detail and edit pages
