+++ docs/documents.md
@@ -13,5 +13,6 @@ The documents feature lets you upload, view, edit, and filter project files.
 
 - Use the filters to search by title, type, or project.
 - View details or download files from each document card.
+- Document detail pages show previews, metadata, and links to edit or download files.
 - Edit document information from the detail page.
 - Delete documents from the detail page; files are removed from storage.
